Unhardcode etherscan contract addresses.
